Update pygamelabeler.py
So much debugging...found a serious problem where I was nulling out ALL the temp box tracking points every time through the loop.  Can't believe that wasn't obvious to me, but it was seriously getting to me.
Now it's displaying a dynamic temp box, crashes when I left click the second time, but we're making progress.  Should be a lot easier from here.