Update pygamelabeler.py
Beautiful - it was a single line dealing with where the Y centroid of the box should be, in the function where I get the screen coords back out of the normalized box values.  My save and load functions are fully working now.
Now to move on to progression through annotating multiple image files in one folder.