Update pygamelabeler.py
Found a simple mistake where I wasn't setting the boxes in memory based on the removeBoxes call.
It's a bit more responsive now that removed boxes are also not displaying.
However, there is still a bug lurking beneath that is a bit more persistent.  There are some cases where boxes are being matched but not removed.  Must be something to do with the file line comparison I'm guessing.  More debugging needed, almost done with this for version 1.0.